# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "AnalysisResponse": {
        "description": "Everything the analysis produces from a series of weigh-ins.",
        "properties": {
          "current": {
            "$ref": "#/components/schemas/CurrentEstimateOut"
          },
          "forecast": {
            "$ref": "#/components/schemas/ForecastOut"
          },
          "meta": {
            "$ref": "#/components/schemas/AnalysisMetaOut"
          },
          "n_obs": {
            "title": "N Obs",
            "type": "integer"
          },
          "observations": {
            "items": {
              "$ref": "#/components/schemas/ObservationOut"
            },
            "title": "Observations",
            "type": "array"
          },
          "params": {
            "$ref": "#/components/schemas/ModelParamsOut"
          },
          "span_days": {
            "title": "Span Days",
            "type": "number"
          },
          "trajectory": {
            "items": {
              "$ref": "#/components/schemas/TrajectoryPointOut"
            },
            "title": "Trajectory",
            "type": "array"
          }
        },
        "required": [
          "n_obs",
          "span_days",
          "params",
          "observations",
          "trajectory",
          "current",
          "forecast",
          "meta"
        ],
        "title": "AnalysisResponse",
        "type": "object"
      },
...
      "DeleteAccountIn": {
        "additionalProperties": false,
        "description": "Confirm irreversible account deletion by retyping the account's email address.",
        "properties": {
          "confirm_email": {
            "description": "Must equal the signed-in account's email address, or the request is rejected.",
            "maxLength": 254,
            "minLength": 3,
            "pattern": "^[^@\\s]+@[^@\\s]+$",
            "title": "Confirm Email",
            "type": "string"
          }
        },
        "required": [
          "confirm_email"
        ],
        "title": "DeleteAccountIn",
        "type": "object"
      },
...
      "ErrorResponse": {
        "description": "The envelope returned for every non-2xx response.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "GoalIn": {
        "additionalProperties": false,
        "description": "Replace the account's goal.\n\nA full replacement, not a merge: an omitted field is stored as ``null``, the same as an\nexplicit one. Never read by any analysis (goal neutrality).",
        "properties": {
          "target_weekly_rate_kg": {
            "anyOf": [
              {
                "maximum": 5.0,
                "minimum": -5.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Target rate of change per week in kilograms, or null for none.",
            "title": "Target Weekly Rate Kg"
          },
          "target_weight_kg": {
            "anyOf": [
              {
                "maximum": 400.0,
                "minimum": 20.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Target weight in kilograms, or null for none.",
            "title": "Target Weight Kg"
          }
        },
        "title": "GoalIn",
        "type": "object"
      },
      "GoalOut": {
        "description": "The user's goal, in kilograms.",
        "properties": {
          "target_weekly_rate_kg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Target rate of change per week, if set. Negative means losing.",
            "title": "Target Weekly Rate Kg"
          },
          "target_weight_kg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Target weight, if set.",
            "title": "Target Weight Kg"
          }
        },
        "required": [
          "target_weight_kg",
          "target_weekly_rate_kg"
        ],
        "title": "GoalOut",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MeOut": {
        "description": "Everything the app needs to know about the signed-in account on load.",
        "properties": {
          "goal": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GoalOut"
              },
              {
                "type": "null"
              }
            ],
            "description": "The user's goal, or null if none is set."
          },
          "measurement_count": {
            "description": "How many measurements are stored.",
            "minimum": 0.0,
            "title": "Measurement Count",
            "type": "integer"
          },
          "preferences": {
            "$ref": "#/components/schemas/PreferencesOut"
          },
          "user": {
            "$ref": "#/components/schemas/UserOut"
          }
        },
        "required": [
          "user",
          "preferences",
          "goal",
          "measurement_count"
        ],
        "title": "MeOut",
        "type": "object"
      },
      "MeasurementBatchIn": {
        "additionalProperties": false,
        "description": "A batch of measurements to store at once, such as a CSV import's accepted rows.",
        "properties": {
          "observations": {
            "description": "Weigh-ins to store, in any order.",
            "items": {
              "$ref": "#/components/schemas/ObservationIn"
            },
            "maxItems": 10000,
            "minItems": 1,
            "title": "Observations",
            "type": "array"
          },
          "source": {
            "default": "manual",
            "description": "Recorded against every measurement created by this batch.",
            "enum": [
              "manual",
              "csv"
            ],
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "observations"
        ],
        "title": "MeasurementBatchIn",
        "type": "object"
      },
      "MeasurementBatchOut": {
        "description": "The outcome of storing a batch.\n\nCarries counts only, not the rows themselves: a caller that wants the resulting history\ncalls ``GET /api/me/measurements`` next, rather than this response duplicating it.",
        "properties": {
          "inserted_count": {
            "description": "Measurements newly stored by this batch.",
            "minimum": 0.0,
            "title": "Inserted Count",
            "type": "integer"
          },
          "skipped_existing_count": {
            "description": "Measurements in this batch that matched one already stored for this account (same instant, same weight once converted to kilograms) and were not duplicated. Re-submitting the same import is therefore idempotent.",
            "minimum": 0.0,
            "title": "Skipped Existing Count",
            "type": "integer"
          }
        },
        "required": [
          "inserted_count",
          "skipped_existing_count"
        ],
        "title": "MeasurementBatchOut",
        "type": "object"
      },
      "MeasurementListOut": {
        "description": "Every measurement the account holds, most recent first.",
        "properties": {
          "count": {
            "description": "How many measurements are in this list.",
            "minimum": 0.0,
            "title": "Count",
            "type": "integer"
          },
          "measurements": {
            "items": {
              "$ref": "#/components/schemas/MeasurementOut"
            },
            "title": "Measurements",
            "type": "array"
          }
        },
        "required": [
          "measurements",
          "count"
        ],
        "title": "MeasurementListOut",
        "type": "object"
      },
      "MeasurementOut": {
        "description": "One stored weigh-in.",
        "properties": {
          "created_at": {
            "description": "When this row was first stored (UTC).",
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "description": "Stable identifier for this measurement.",
            "title": "Id",
            "type": "string"
          },
          "source": {
            "description": "How this measurement was added.",
            "enum": [
              "manual",
              "csv"
            ],
            "title": "Source",
            "type": "string"
          },
          "timestamp": {
            "description": "When the weigh-in happened, in UTC.",
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          },
          "unit": {
            "description": "The unit the weight was entered in.",
            "enum": [
              "kg",
              "lb"
            ],
            "title": "Unit",
            "type": "string"
          },
          "updated_at": {
            "description": "When this row was last changed (UTC).",
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "weight": {
            "description": "The weight as entered, in the unit field.",
            "title": "Weight",
            "type": "number"
          }
        },
        "required": [
          "id",
          "timestamp",
          "weight",
          "unit",
          "source",
          "created_at",
          "updated_at"
        ],
        "title": "MeasurementOut",
        "type": "object"
      },
...
      "ObservationIn": {
        "additionalProperties": false,
        "description": "One submitted weight measurement.",
        "properties": {
          "timestamp": {
            "description": "ISO-8601 instant of the weigh-in. Must carry a timezone offset, and must not be so late that a 90-day forecast from it is unrepresentable.",
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          },
          "unit": {
            "default": "kg",
            "description": "Unit of the weight field.",
            "enum": [
              "kg",
              "lb"
            ],
            "title": "Unit",
            "type": "string"
          },
          "weight": {
            "description": "The measured value, in the unit given by the unit field.",
            "exclusiveMinimum": 0.0,
            "title": "Weight",
            "type": "number"
          }
        },
        "required": [
          "timestamp",
          "weight"
        ],
        "title": "ObservationIn",
        "type": "object"
      },
...
      "PreferencesIn": {
        "additionalProperties": false,
        "description": "Replace the account's display preferences.",
        "properties": {
          "display_unit": {
            "description": "Unit weights should be displayed in.",
            "enum": [
              "kg",
              "lb"
            ],
            "title": "Display Unit",
            "type": "string"
          }
        },
        "required": [
          "display_unit"
        ],
        "title": "PreferencesIn",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Estimates the underlying weight trajectory behind noisy scale readings, quantifies how\nuncertain that estimate is, and forecasts where it is heading.\n\nIntervals describe the latent weight rather than a future scale reading, and the trajectory\nis the online (filtered) estimate, so no point on it changes retroactively.\n\nNot a medical device: this estimates and forecasts a measurement trend. It does not\ndiagnose, treat or prescribe.",
    "summary": "Probabilistic weight-trend estimation and forecasting.",
    "title": "HealthTrend",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/me": {
      "delete": {
        "description": "Hard-delete the account and everything it owns.\n\nIts sessions, measurements, preferences, goal and issued sign-in codes. Requires\nretyping the account's own email address as confirmation.",
        "operationId": "delete_my_account_api_me_delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteAccountIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "confirm_email did not match the account."
          }
        },
        "summary": "Permanently delete the account and everything it owns",
        "tags": [
          "account"
        ]
      },
      "get": {
        "description": "Return the signed-in user, their preferences, their goal and their measurement count.",
        "operationId": "me_api_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not signed in."
          }
        },
        "summary": "Describe the signed-in account",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/me/analysis": {
      "get": {
        "description": "Analyse every measurement the account holds, forecasting from the current instant.\n\nCalls the same :func:`app.services.analysis.analyse_submitted` a submitted series does;\n``meta.source`` reads ``\"account\"`` in place of ``\"submitted\"`` and nothing else differs.",
        "operationId": "analyse_my_account_api_me_analysis_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "No measurements are stored, or one is invalid."
          }
        },
        "summary": "Analyse the account's own stored measurement history",
        "tags": [
          "account"
        ]
      }
    },
...
    "/api/me/goal": {
      "delete": {
        "description": "Remove the account's goal, if one is set. Not an error if none is.",
        "operationId": "delete_my_goal_api_me_goal_delete",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "summary": "Remove the account's goal",
        "tags": [
          "account"
        ]
      },
      "put": {
        "description": "Replace the account's goal. Never read by any analysis (goal neutrality).",
        "operationId": "update_my_goal_api_me_goal_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoalIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Replace the account's goal",
        "tags": [
          "account"
        ]
      }
    },
    "/api/me/measurements": {
      "get": {
        "description": "Return every measurement the account holds, most recent first.",
        "operationId": "list_my_measurements_api_me_measurements_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeasurementListOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List every stored measurement",
        "tags": [
          "account"
        ]
      },
      "post": {
        "description": "Store one manually entered measurement.",
        "operationId": "create_my_measurement_api_me_measurements_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ObservationIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeasurementOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The account's measurement cap was reached."
          }
        },
        "summary": "Store one measurement",
        "tags": [
          "account"
        ]
      }
    },
    "/api/me/measurements/batch": {
      "post": {
        "description": "Store a batch of measurements. Re-submitting the same batch inserts nothing new.",
        "operationId": "batch_create_my_measurements_api_me_measurements_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MeasurementBatchIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeasurementBatchOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The account's measurement cap was reached."
          }
        },
        "summary": "Store several measurements at once, such as a CSV import's accepted rows",
        "tags": [
          "account"
        ]
      }
    },
    "/api/me/measurements/{measurement_id}": {
      "delete": {
        "description": "Delete a stored measurement.",
        "operationId": "delete_my_measurement_api_me_measurements__measurement_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "measurement_id",
            "required": true,
            "schema": {
              "title": "Measurement Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "No measurement exists under that id."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete a stored measurement",
        "tags": [
          "account"
        ]
      },
      "put": {
        "description": "Replace a stored measurement's timestamp, weight and unit.\n\n``measurement_id`` belonging to another account is rejected identically to one that does\nnot exist: both raise ``measurement_not_found``.",
        "operationId": "update_my_measurement_api_me_measurements__measurement_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "measurement_id",
            "required": true,
            "schema": {
              "title": "Measurement Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ObservationIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeasurementOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "No measurement exists under that id."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Replace a stored measurement",
        "tags": [
          "account"
        ]
      }
    },
    "/api/me/preferences": {
      "put": {
        "description": "Replace the account's display preferences.",
        "operationId": "update_my_preferences_api_me_preferences_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreferencesIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferencesOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Replace the account's display preferences",
        "tags": [
          "account"
        ]
      }
    },
...
  }
}